requirements.txt
Flask==2.1.2
requests==2.28.1
psycopg2-binary==2.9.3
apscheduler==3.9.1.post1
plotly==5.7.0
pandas
elasticsearch